Renamed src/hooks to src/services to reduce confusion
