+++ src/index.js
@@ -1,13 +1,13 @@
 
 module.exports = Object.assign({},
-  require('./hooks/alter-data'),
-  require('./hooks/query-params'),
-  require('./hooks/populate'),
-  require('./hooks/legacyPopulate'),
-  require('./hooks/serialize'),
-  require('./hooks/database'),
-  require('./hooks/validate'),
-  require('./hooks/utilities'),
-  require('./hooks/conditionals'),
-  require('./hooks/utils'),
+  require('./services/alter-data'),
+  require('./services/query-params'),
+  require('./services/populate'),
+  require('./services/legacyPopulate'),
+  require('./services/serialize'),
+  require('./services/database'),
+  require('./services/validate'),
+  require('./services/utilities'),
+  require('./services/conditionals'),
+  require('./services/utils'),
 );
