+++ test/hooks/utilities/promiseToCallback.test.js
@@ -1,6 +1,6 @@
 
 import { assert } from 'chai';
-import { promiseToCallback } from '../../../src/hooks';
+import { promiseToCallback } from '../../../src/services';
 
 const promise1 = (ifResolve) => new Promise((resolve, reject) => {
   return ifResolve ? resolve('ok') : reject('bad');
