+++ test/hooks/validate/validate.test.js
@@ -1,6 +1,6 @@
 
 import { assert } from 'chai';
-import hooks from '../../../src/hooks';
+import hooks from '../../../src/services';
 import errors from 'feathers-errors';
 
 var fcnSync;
